Prepared the help function for printing information on loaded filters too.
For each filter a header is printed containing its name and if available its description. The headline also contains a note about prefixing the name of the filter options by "filter-" if they are identical to an option of aspell itselve. This header then is followed as before by the options available for the entire filter.
Before each of these filter descriptive parts an empty line is printed to optically separate from the list before.

[Imported from GNU Aspell CVS Repository.]