Changed version from 0.51 to 0.60.
Changed the format of the character data file.

Enhanced the format of the dictionary so that case, accent, and length
information is stored with a word to speed up checking.

Added appendix to manual giving an overview of languages which Aspell should
eventually be able to support.

Other changes I probally forgot about.

[Imported from GNU Aspell CVS Repository.]